pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "projective-consistency"
version = "0.1.0"
description = "Baseline evaluator backbone for projective consistency in geometry-rich generated images."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" },
]
dependencies = [
  "numpy>=1.26,<3",
  "Pillow>=10.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
opencv = [
  "opencv-python>=4.10.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

